Prevent propagation of window parameters when switching perspectives
`current-window-configuration` and `set-current-window-configuration` do not
care about window parameters so ensure that they are not in any of the windows
when creating new perspectives.

This is important because certain commands like `delete-other-windows` and
`split-window` take into account window parameters like `window-side` that
may inhibit these commands. We may want this behavior on the current
perspective, but not in any windows on the new perspective.