Generalizes sorting of perspectives and adds sort-by-creation-time.
This is valuable to me because I often use perspectives during yak
shaves and other multi-project tasks where I need to "push" and "pop"
things to work on. By sorting by persp-time-created, this simulates
that perfectly:

+ While working on a task, find a tangent I need to go off on.
+ Create a new perspective.
+ Work on tangent. Possibly going back and forth between multiple perspectives.
+ When done with tangent, kill the perspective, going back to next-most-recent.

This switches from `persp-sort-chronologically` (bool) to
`persp-sort` (symbol). There might be some smoother way to honor the
old value and auto-convert to the new variable, but I don't know how.

This also adds `created-time` to the perspective struct. Loading this
code in your current emacs breaks everything. I don't know if it is
possible to migrate existing structs or otherwise patch this code to
work with old structs. Restarting emacs seems fine but I'm open to
suggestion.