Fix: persp-switch-buffer use ivy-switch-buffer as caller
Using counsel-switch-buffer was causing issues with the
all-the-icons-ivy integration, as well as preventing all of the ivy
actions from showing up in the hydra menu. for example, killing marked
buffers is not present in the actions if counsel-switch-buffer is
used, but it does appear when the caller is ivy-switch-buffer

this may cause issues for people that have further customization on
the counsel-switch-buffer caller, but from what i can tell, the
counsel-switch-buffer function doesn't even pass itself down as the
caller in its implementation.