persp-get-buffers: get a perspective's list of buffers
New utility function to return a perspective's list of buffers for a
specific frame or for the selected one.  As argument, accepts a name
or a proper perspective data, if nil or if not given, default to the
current perspective.  If a frame is nil or not given, default to the
selected frame.