basic-persp-killing-buffers: test killing buffers
A buffer should not be killable when it's a perspective's last left
buffer.  This prevents adding a buffer from another perspective.  A
perspective can be killed, though, even if it has only one buffer.

It should be possible to remove buffers from perspectives, unless a
buffer is the perspective's last left buffer.