basic-persp-test-buffer-in-persps: test new ert's function
Test the new utility function 'persp-test-buffer-in-persps'.

As arguments, expect a buffer or it's name, and perspective names, or
perspective data (aka 'perspective-p' objects).  The buffer must be a
live buffer, but a perspective data not necessarily should reflect an
existing perspective.

The buffer must be found exactly in the given perspectives and not in
all other existing perspectives.

Perspectives owning the buffer may be provided as `perspective-p' data
passed as arguments, even when it isn't of an existing perspective.