persp-remove-buffer: bury buffer walking windows in one way loop
A one way loop, with pre-defined start and end, prevents any infinite
ping-pong.

This is not just syntactic sugar.  In a "while buried buffer found in
window" loop, there is the risk that when the buffer is buried in one
window, it may be un-buried in another (for "mystic" reasons) and the
loop will never end.