persp-activate: forget windows buffers not in perspective
Implement a dirty flag to let a perspective forget buffers which are
found in windows, but do not belong to the perspective.

A buffer removed from a perspective by manipulating the frame's hash
table, requires the perspective's dirty flag to be set.

This is required since 'persp-maybe-kill-buffer' no longer updates a
perspective's windows configuration.  When a removed buffer is still
part of a windows configuration, a perspective will get back removed
buffers when re-activated.